Changing sprite to train
The snake image is changed to train. There are multiple colored images for the train's body, so the code was changed to make the colors move around and select different colors. Also when the train moves up or down, the orientation of the train's head remains in the same direction it was going in.